+++ PanteraCRM/Negocios/almacenNE.cs
@@ -48,7 +48,7 @@ namespace Negocios
         {
             return almacenDL.SaldoAlmacenAdiconar( a1,  p1,  cantidad);
         }
-        public static int CambiarSaldoComprometido(int a1, int p1, decimal cantidad)
+        public static int CambiarSaldoComprometido(int a1, int p1, int cantidad)
         {
             return almacenDL.CambiarSaldoComprometido(a1, p1, cantidad);
         }
